requirements.txt
Flask
pillow
numpy
opencv-python
reportlab